requirements.txt
# ComfyUI-PrimitiveAnything requirements

# Core dependencies
trimesh>=4.0.0
numpy>=1.22.0
scipy>=1.7.0
seaborn>=0.12.0
pyyaml>=6.0

# HuggingFace for model downloading
huggingface_hub
accelerate>=0.20.0

# 3D processing
open3d>=0.17.0
scikit-image>=0.20.0
mesh2sdf
mesh-to-sdf
opencv-python>=4.5.0

# PrimitiveAnything model dependencies
beartype
einops
gateloop_transformer
pytorch_custom_utils
x_transformers
toolz
environs
jaxtyping
omegaconf
transformers
pytorch_lightning
